requirements.txt
playwright
python-dotenv
gspread
google-auth
pandas
openpyxl
xlrd
lxml
